- Split utils.py into utils.py and parsing.py - Start working on tests - Allow keyword arguments to BIDSReport.generate that are fed into BIDSLayout getters. - Fix some issues with how config dicts are loaded. - Change utils.warnings to utils.reminder
