+++ bids/reports/config/converters.json
@@ -1,28 +1,28 @@
 {
     "dir": {
-        "i": "left to right", 
-        "i-": "right to left", 
-        "j": "posterior to anterior", 
-        "j-": "anterior to posterior", 
-        "k": "inferior to superior", 
+        "i": "left to right",
+        "i-": "right to left",
+        "j": "posterior to anterior",
+        "j-": "anterior to posterior",
+        "k": "inferior to superior",
         "k-": "superior to inferior"
-    }, 
+    },
     "seq": {
-        "EP": "echo planar", 
-        "GR": "gradient recalled", 
-        "IR": "inversion recovery", 
-        "RM": "research mode", 
+        "EP": "echo planar",
+        "GR": "gradient recalled",
+        "IR": "inversion recovery",
+        "RM": "research mode",
         "SE": "spin echo"
-    }, 
+    },
     "seqvar": {
-        "MP": "MAG prepared", 
-        "MTC": "magnetization transfer contrast", 
-        "NONE": "no sequence variant", 
-        "OSP": "oversampling phase", 
-        "SK": "segmented k-space", 
-        "SP": "spoiled", 
-        "SS": "steady state", 
+        "MP": "MAG prepared",
+        "MTC": "magnetization transfer contrast",
+        "NONE": "no sequence variant",
+        "OSP": "oversampling phase",
+        "SK": "segmented k-space",
+        "SP": "spoiled",
+        "SS": "steady state",
         "TRSS": "time reversed steady state"
-    }, 
-    "task": {}
+    },
+    "task": {"rest": "resting state"}
 }
